# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Content Workflow API
  description: |
    REST API for AI-powered content creation, translation, and review workflows.
    
    This API provides comprehensive endpoints for managing marketing campaigns,
    content pieces, AI generation, translations, and review processes.
    
    **Features:**
    - Campaign and content management
    - Multi-model AI content generation (OpenAI, Anthropic)
    - Automated translation services
    - Content analysis and sentiment detection
    - Review and approval workflows
    - Real-time monitoring and health checks
    
  version: 1.0.0
  contact:
    name: AI Content Workflow Team
    url: https://github.com/your-org/ai-content-workflow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.example.com/v1
    description: Production server
paths:
...
  /campaigns:
    get:
      summary: List all campaigns
      description: Retrieve a paginated list of all campaigns with optional filtering
      operationId: getCampaigns
      tags:
        - Campaigns
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Filter by campaign status
          schema:
            type: string
            enum: [active, inactive, completed]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Campaign'
                  total:
                    type: integer
                  page:
                    type: integer
                  totalPages:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      summary: Create a new campaign
      description: Create a new marketing campaign
      operationId: createCampaign
      tags:
        - Campaigns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Campaign created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /campaigns/{id}:
    get:
      summary: Get campaign by ID
      description: Retrieve a specific campaign with all related content pieces
      operationId: getCampaignById
      tags:
        - Campaigns
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      summary: Update campaign
      description: Update an existing campaign
      operationId: updateCampaign
      tags:
        - Campaigns
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCampaignRequest'
      responses:
        '200':
          description: Campaign updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      summary: Delete campaign
      description: Delete a campaign and all associated content
      operationId: deleteCampaign
      tags:
        - Campaigns
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Campaign deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Content endpoints
...
components:
  schemas:
    Campaign:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the campaign
        name:
          type: string
          description: Campaign name
          example: "Summer Sale 2024"
        description:
          type: string
          nullable: true
          description: Campaign description
          example: "Marketing campaign for summer products"
        status:
          type: string
          enum: [active, inactive, completed]
          description: Campaign status
        defaultLanguage:
          type: string
          description: Default language code
          example: "en"
        targetLanguages:
          type: array
          items:
            type: string
          description: List of target language codes
          example: ["es", "fr", "de"]
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        contentPieces:
          type: array
          items:
            $ref: '#/components/schemas/ContentPiece'
          description: Associated content pieces
    ContentPiece:
      type: object
      properties:
        id:
          type: integer
          format: int64
        campaignId:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/ContentType'
        originalContent:
          type: string
          nullable: true
          description: Original content text
        language:
          type: string
          description: Content language code
          example: "en"
        status:
          $ref: '#/components/schemas/ContentStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        aiGenerations:
          type: array
          items:
            $ref: '#/components/schemas/AIGeneration'
        translations:
          type: array
          items:
            $ref: '#/components/schemas/Translation'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
    AIGeneration:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contentPieceId:
          type: integer
          format: int64
        generatedText:
          type: string
          description: AI-generated content
        promptUsed:
          type: string
          nullable: true
          description: Prompt used for generation
        modelVersion:
          type: string
          description: AI model version used
          example: "gpt-4"
        metadata:
          type: string
          nullable: true
          description: JSON metadata including analysis data
        createdAt:
          type: string
          format: date-time
    Translation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contentPieceId:
          type: integer
          format: int64
        targetLanguage:
          type: string
          description: Target language code
          example: "es"
        translatedText:
          type: string
          description: Translated content
        status:
          type: string
          enum: [draft, completed]
        qualityScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
          nullable: true
          description: Translation quality score
        createdAt:
          type: string
          format: date-time
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contentPieceId:
          type: integer
          format: int64
        language:
          type: string
          description: Language of the reviewed content
        status:
          type: string
          enum: [approved, rejected, needs_changes]
        feedback:
          type: string
          nullable: true
          description: Reviewer feedback
        reviewerName:
          type: string
          nullable: true
          description: Name of the reviewer
        createdAt:
          type: string
          format: date-time
    ContentType:
      type: string
      enum: [headline, description, blog_post, social_media, email]
      description: Type of content piece
    ContentStatus:
      type: string
      enum: [draft, ai_generated, under_review, approved, rejected]
      description: Status of content piece
    
    # Request schemas
    CreateCampaignRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "Summer Sale 2024"
        description:
          type: string
          maxLength: 1000
          example: "Marketing campaign for summer products"
        defaultLanguage:
          type: string
          default: "en"
          example: "en"
        targetLanguages:
          type: array
          items:
            type: string
          example: ["es", "fr", "de"]
    UpdateCampaignRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        status:
          type: string
          enum: [active, inactive, completed]
        targetLanguages:
          type: array
          items:
            type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid request data"
            code: "BAD_REQUEST"
            timestamp: "2024-01-01T12:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
            timestamp: "2024-01-01T12:00:00Z"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error occurred"
            code: "INTERNAL_ERROR"
            timestamp: "2024-01-01T12:00:00Z"
...